requirements.txt
boto3
openai
load-dotenv
streamlit
moviepy
imageio
opencv-python
streamlit-option-menu